Rekall Profile Generator
Rekall currently lacks the ability to generate/convert profiles for Linux
compiled for Aarch64. Thus, we introduce the Rekall Profile Generator that is
capable of generating such Rekall profiles. The Rekall profile format comprises
the following fields in JSON format:

{
  "$CONFIG": {},
  "$CONSTANTS": {},
  "$ENUMS": {},
  "$METADATA": {},
  "$REVENUMS": {},
  "$STRUCTS": {}
}

The current implementation of the Rekall Profile Generator parses the Linux
config ("$CONFIG") file and the System.map ("$CONSTANTS") and fills the
corresponding entries in the Rekall profile.